State and VM refactoring
