+++ evm-tests/jsontests/Cargo.toml
@@ -22,17 +22,19 @@ sha3.workspace = true
 serde.workspace = true
 
 bytecount = "0.6"
-ethereum = "0.15.0"
+# ethereum = "0.15.0"
 serde_json = "1.0"
 hex = "0.4"
 clap = { version = "4.5", features = ["cargo"] }
-ethjson = { path = "../ethjson", features = ["test-helpers"] }
 libsecp256k1 = "0.7"
-ethcore-builtin = { path = "../ethcore-builtin" }
 parity-bytes = "0.1"
 hex-literal = "0.4"
+
+ethjson = { path = "../ethjson", features = ["test-helpers"] }
+ethcore-builtin = { path = "../ethcore-builtin" }
 
 [features]
 enable-slow-tests = []
 print-debug = ["aurora-evm/print-debug"]
+
 dump-state = ["aurora-evm/with-serde"]
