+++ evm-tests/jsontests/src/state.rs
@@ -1,3 +1,7 @@
+use crate::execution_results::{FailedTestDetails, TestExecutionResult};
+use crate::state_dump::StateTestsDump;
+use crate::types::spec::Spec;
+use crate::types::StateTestCase;
 use crate::utils::eip_4844;
 use crate::utils::transaction::InvalidTxReason;
 use aurora_evm::backend::{ApplyBackend, MemoryAccount, MemoryBackend, MemoryVicinity};
@@ -7,11 +11,7 @@ use aurora_evm::executor::stack::{
 };
 use aurora_evm::utils::U64_MAX;
 use aurora_evm::{Config, Context, ExitError, ExitReason, ExitSucceed};
-use ethjson::hash::Address;
-use ethjson::spec::builtin::{AltBn128ConstOperations, AltBn128Pairing, PricingAt};
-use ethjson::spec::{ForkSpec, Pricing};
-use ethjson::test_helpers::state::PostStateResult;
-use ethjson::uint::Uint;
+
 use libsecp256k1::SecretKey;
 use primitive_types::{H160, H256, U256};
 use serde::Deserialize;
@@ -20,145 +20,6 @@ use std::collections::BTreeMap;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::{Arc, LazyLock};
-
-#[derive(Default, Debug, Clone)]
-#[allow(clippy::struct_excessive_bools)]
-pub struct VerboseOutput {
-    pub verbose: bool,
-    pub verbose_failed: bool,
-    pub very_verbose: bool,
-    pub print_state: bool,
-}
-
-#[derive(Default, Debug, Clone)]
-#[cfg_attr(feature = "dump-state", derive(serde::Serialize, serde::Deserialize))]
-pub struct StateTestsDump {
-    pub state: BTreeMap<H160, MemoryAccount>,
-    pub caller: H160,
-    pub gas_price: U256,
-    pub effective_gas_price: U256,
-    pub caller_secret_key: H256,
-    pub used_gas: u64,
-    pub state_hash: H256,
-    pub result_state: BTreeMap<H160, MemoryAccount>,
-    pub to: H160,
-    pub value: U256,
-    pub data: Vec<u8>,
-    pub gas_limit: u64,
-    pub access_list: Vec<(H160, Vec<H256>)>,
-}
-
-trait StateTestsDumper {
-    fn set_state(&mut self, _state: &BTreeMap<H160, MemoryAccount>) {}
-    fn set_used_gas(&mut self, _used_gas: u64) {}
-    fn set_vicinity(&mut self, _vicinity: &MemoryVicinity) {}
-    fn set_tx_data(
-        &mut self,
-        _to: H160,
-        _value: U256,
-        _data: Vec<u8>,
-        _gas_limit: u64,
-        _access_list: Vec<(H160, Vec<H256>)>,
-    ) {
-    }
-    fn set_caller_secret_key(&mut self, _caller_secret_key: H256) {}
-    fn set_state_hash(&mut self, _state_hash: H256) {}
-    fn set_result_state(&mut self, _state: &BTreeMap<H160, MemoryAccount>) {}
-    fn dump_to_file(&self, _spec: &ForkSpec) {}
-}
-
-#[cfg(not(feature = "dump-state"))]
-impl StateTestsDumper for StateTestsDump {}
-
-#[cfg(feature = "dump-state")]
-impl StateTestsDumper for StateTestsDump {
-    fn set_state(&mut self, state: &BTreeMap<H160, MemoryAccount>) {
-        self.state = state.clone();
-    }
-
-    fn set_used_gas(&mut self, used_gas: u64) {
-        self.used_gas = used_gas;
-    }
-
-    fn set_vicinity(&mut self, vicinity: &MemoryVicinity) {
-        self.caller = vicinity.origin;
-        self.gas_price = vicinity.gas_price;
-        self.effective_gas_price = vicinity.effective_gas_price;
-    }
-
-    fn set_tx_data(
-        &mut self,
-        to: H160,
-        value: U256,
-        data: Vec<u8>,
-        gas_limit: u64,
-        access_list: Vec<(H160, Vec<H256>)>,
-    ) {
-        self.to = to;
-        self.value = value;
-        self.data = data;
-        self.gas_limit = gas_limit;
-        self.access_list = access_list;
-    }
-
-    fn set_caller_secret_key(&mut self, caller_secret_key: H256) {
-        self.caller_secret_key = caller_secret_key;
-    }
-
-    fn set_state_hash(&mut self, state_hash: H256) {
-        self.state_hash = state_hash;
-    }
-
-    fn set_result_state(&mut self, state: &BTreeMap<H160, MemoryAccount>) {
-        self.result_state = state.clone();
-    }
-
-    fn dump_to_file(&self, spec: &ForkSpec) {
-        use std::time::{SystemTime, UNIX_EPOCH};
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_micros();
-        let path = format!("{spec:?}_BLS12-382_{now}.json");
-        let json = serde_json::to_string(&self).unwrap();
-        std::fs::write(path, json).unwrap();
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct FailedTestDetails {
-    pub name: String,
-    pub spec: ForkSpec,
-    pub index: usize,
-    pub expected_hash: H256,
-    pub actual_hash: H256,
-    pub state: BTreeMap<H160, MemoryAccount>,
-}
-
-#[derive(Clone, Debug)]
-pub struct TestExecutionResult {
-    pub total: u64,
-    pub failed: u64,
-    pub failed_tests: Vec<FailedTestDetails>,
-}
-
-impl TestExecutionResult {
-    #[allow(clippy::new_without_default)]
-    #[must_use]
-    pub const fn new() -> Self {
-        Self {
-            total: 0,
-            failed: 0,
-            failed_tests: Vec::new(),
-        }
-    }
-
-    pub fn merge(&mut self, src: Self) {
-        self.failed_tests.extend(src.failed_tests);
-        self.total += src.total;
-        self.failed += src.failed;
-    }
-}
 
 #[derive(Deserialize, Debug)]
 pub struct Test(ethjson::test_helpers::state::State);
@@ -280,7 +141,10 @@ impl Test {
     }
 }
 
-type LazyPrecompiles = LazyLock<BTreeMap<H160, ethcore_builtin::Builtin>>;
+#[derive(Debug, Clone)]
+struct Precompile;
+
+type LazyPrecompiles = LazyLock<BTreeMap<H160, Precompile>>;
 static ISTANBUL_BUILTINS: LazyPrecompiles = LazyLock::new(istanbul_builtins);
 static BERLIN_BUILTINS: LazyPrecompiles = LazyLock::new(berlin_builtins);
 static CANCUN_BUILTINS: LazyPrecompiles = LazyLock::new(cancun_builtins);
@@ -409,7 +273,7 @@ impl JsonPrecompile {
 }
 
 #[allow(clippy::too_many_lines)]
-fn istanbul_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
+fn istanbul_builtins() -> BTreeMap<H160, Precompile> {
     use ethjson::spec::builtin::{BuiltinCompat, Linear, Modexp, PricingCompat};
 
     let builtins: BTreeMap<Address, BuiltinCompat> = BTreeMap::from([
@@ -522,19 +386,12 @@ fn istanbul_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
     ]);
     builtins
         .into_iter()
-        .map(|(address, builtin)| {
-            (
-                address.into(),
-                ethjson::spec::Builtin::from(builtin).try_into().unwrap(),
-            )
-        })
+        .map(|(address, builtin)| (address.into(), Precompile))
         .collect()
 }
 
 #[allow(clippy::too_many_lines)]
-fn berlin_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
-    use ethjson::spec::builtin::{BuiltinCompat, Linear, Modexp, PricingCompat};
-
+fn berlin_builtins() -> BTreeMap<H160, Precompile> {
     let builtins: BTreeMap<Address, BuiltinCompat> = BTreeMap::from([
         (
             Address(H160::from_low_u64_be(1)),
@@ -654,9 +511,7 @@ fn berlin_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
         .collect()
 }
 
-fn cancun_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
-    use ethjson::spec::builtin::{BuiltinCompat, Linear, PricingCompat};
-
+fn cancun_builtins() -> BTreeMap<H160, Precompile> {
     let mut builtins = berlin_builtins();
     builtins.insert(
         Address(H160::from_low_u64_be(0xA)).into(),
@@ -674,9 +529,7 @@ fn cancun_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
     builtins
 }
 
-fn prague_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
-    use ethjson::spec::builtin::{BuiltinCompat, Linear, PricingCompat};
-
+fn prague_builtins() -> BTreeMap<H160, Precompile> {
     let mut builtins = cancun_builtins();
     builtins.insert(
         Address(H160::from_low_u64_be(0xB)).into(),
@@ -764,31 +617,15 @@ fn prague_builtins() -> BTreeMap<H160, ethcore_builtin::Builtin> {
 ///
 /// This function will panic if thread spawning or joining fails.
 #[must_use]
-pub fn test(
-    verbose_output: VerboseOutput,
-    name: &str,
-    test: Test,
-    specific_spec: Option<ForkSpec>,
-    file_name: Arc<PathBuf>,
-) -> TestExecutionResult {
+pub fn test(test_config: TestConfig, name: String, test: StateTestCase) -> TestExecutionResult {
     use std::thread;
 
     const STACK_SIZE: usize = 16 * 1024 * 1024;
 
-    let name = name.to_string();
     // Spawn thread with explicit stack size
-
     let child = thread::Builder::new()
         .stack_size(STACK_SIZE)
-        .spawn(move || {
-            test_run(
-                &verbose_output,
-                &name,
-                &test,
-                specific_spec.as_ref(),
-                &file_name,
-            )
-        })
+        .spawn(move || test_run(&test_config, &name, &test))
         .unwrap();
 
     // Wait for thread to join
@@ -1210,346 +1047,342 @@ fn check_validate_exit_reason(
 }
 
 #[allow(clippy::cognitive_complexity, clippy::too_many_lines)]
-fn test_run(
-    verbose_output: &VerboseOutput,
-    name: &str,
-    test: &Test,
-    specific_spec: Option<&ForkSpec>,
-    file_name: &Arc<PathBuf>,
-) -> TestExecutionResult {
-    let mut tests_result = TestExecutionResult::new();
-    let test_tx = &test.0.transaction;
-    for (spec, states) in &test.0.post_states {
-        // Run tests for specific SPEC (Hard fork)
-        if let Some(s) = specific_spec.as_ref() {
-            if *s != spec {
-                continue;
-            }
-        }
-
-        let (gasometer_config, delete_empty) = match spec {
-            ForkSpec::Istanbul => (Config::istanbul(), true),
-            ForkSpec::Berlin => (Config::berlin(), true),
-            ForkSpec::London => (Config::london(), true),
-            ForkSpec::Merge | ForkSpec::Paris => (Config::merge(), true),
-            ForkSpec::Shanghai => (Config::shanghai(), true),
-            ForkSpec::Cancun => (Config::cancun(), true),
-            ForkSpec::Prague => (Config::prague(), true),
-            _ => {
-                continue;
-            }
-        };
-
-        // EIP-4844
-        let blob_gas_price = test
-            .0
-            .env
-            .current_excess_blob_gas
-            .map(|current_excess_blob_gas| {
-                eip_4844::calc_blob_gas_price(current_excess_blob_gas.0.as_u64())
-            })
-            .or_else(|| {
-                test.0
-                    .env
-                    .parent_blob_gas_used
-                    .zip(test.0.env.parent_excess_blob_gas)
-                    .map(|(parent_blob_gas_used, parent_excess_blob_gas)| {
-                        let excess_blob_gas = eip_4844::calc_excess_blob_gas(
-                            parent_blob_gas_used.0.as_u64(),
-                            parent_excess_blob_gas.0.as_u64(),
-                        );
-                        eip_4844::calc_blob_gas_price(excess_blob_gas)
-                    })
-            });
-        // EIP-4844
-        let data_max_fee = if gasometer_config.has_shard_blob_transactions {
-            let max_fee_per_blob_gas = test_tx.max_fee_per_blob_gas.unwrap_or_default().0;
-            Some(eip_4844::calc_max_data_fee(
-                max_fee_per_blob_gas,
-                test_tx.blob_versioned_hashes.len(),
-            ))
-        } else {
-            None
-        };
-        let data_fee = if gasometer_config.has_shard_blob_transactions {
-            Some(eip_4844::calc_data_fee(
-                blob_gas_price.expect("expect blob_gas_price"),
-                test_tx.blob_versioned_hashes.len(),
-            ))
-        } else {
-            None
-        };
-
-        let original_state = test.unwrap_to_pre_state();
-        let vicinity = test.unwrap_to_vicinity(spec, blob_gas_price);
-        if let Err(tx_err) = vicinity {
-            tests_result.total += states.len() as u64;
-            let h = states.first().unwrap().hash.0;
-            // if vicinity could not be computed then the transaction was invalid so we simply
-            // check the original state and move on
-            let (is_valid_hash, actual_hash) = crate::utils::check_valid_hash(&h, &original_state);
-            if !is_valid_hash {
-                tests_result.failed_tests.push(FailedTestDetails {
-                    expected_hash: h,
-                    actual_hash,
-                    index: 0,
-                    name: String::from_str(name).unwrap(),
-                    spec: spec.clone(),
-                    state: original_state,
-                });
-                if verbose_output.verbose_failed {
-                    println!(" [{spec:?}] {name}: {tx_err:?} ... validation failed\t<----");
-                }
-                tests_result.failed += 1;
-                continue;
-            }
-            assert_vicinity_validation(&tx_err, states, spec, name, &file_name.clone());
-            // As it's expected validation error - skip the test run
-            continue;
-        }
-        let vicinity = vicinity.unwrap();
-        let caller = test.unwrap_caller();
-        let caller_balance = original_state
-            .get(&caller)
-            .map_or_else(U256::zero, |acc| acc.balance);
-        // EIP-3607
-        let caller_code = original_state
-            .get(&caller)
-            .map_or_else(Vec::new, |acc| acc.code.clone());
-        // EIP-7702 - check if it's delegated designation. If it's delegation designation then
-        // even if `caller_code` is non-empty transaction should be executed.
-        let is_delegated = original_state
-            .get(&caller)
-            .is_some_and(|c| Authorization::is_delegated(&c.code));
-
-        for (i, state) in states.iter().enumerate() {
-            let transaction = test_tx.select(&state.indexes);
-            let mut backend = MemoryBackend::new(&vicinity, original_state.clone());
-            tests_result.total += 1;
-            // Test case may be expected to fail with an unsupported tx type if the current fork is
-            // older than Berlin (see EIP-2718). However, this is not implemented in sputnik itself and rather
-            // in the code hosting sputnik. https://github.com/rust-blockchain/evm/pull/40
-            let expect_tx_type_not_supported = matches!(
-                spec,
-                ForkSpec::EIP150
-                    | ForkSpec::EIP158
-                    | ForkSpec::Frontier
-                    | ForkSpec::Homestead
-                    | ForkSpec::Byzantium
-                    | ForkSpec::Constantinople
-                    | ForkSpec::ConstantinopleFix
-                    | ForkSpec::Istanbul
-                    | ForkSpec::Berlin
-            ) && TxType::from_txbytes(&state.txbytes)
-                != TxType::Legacy
-                && state.expect_exception.as_deref() == Some("TR_TypeNotSupported");
-            if expect_tx_type_not_supported {
-                continue;
-            }
-
-            let gas_limit: u64 = transaction.gas_limit.into();
-            let data: Vec<u8> = transaction.data.clone().into();
-            let valid_tx = crate::utils::transaction::validate(
-                &transaction,
-                test.0.env.gas_limit.0,
-                caller_balance,
-                &gasometer_config,
-                test_tx,
-                &vicinity,
-                blob_gas_price,
-                data_max_fee,
-                spec,
-                state,
-            );
-            // Only execute valid transactions
-            if let Err(err) = &valid_tx {
-                if check_validate_exit_reason(err, state.expect_exception.as_ref(), name, spec) {
+fn test_run(_test_config: &TestConfig, _name: &str, test: &StateTestCase) -> TestExecutionResult {
+    let tests_result = TestExecutionResult::new();
+    let _test_tx = &test.transaction;
+    /*
+        for (spec, states) in &test.post_states {
+            // Run tests for specific SPEC (Hard fork)
+            if let Some(s) = test_config.spec.as_ref() {
+                if s != spec {
                     continue;
                 }
             }
-            let authorization_list = valid_tx.unwrap();
 
-            // We do not check overflow after TX validation
-            let total_fee = if let Some(data_fee) = data_fee {
-                vicinity.effective_gas_price * gas_limit + data_fee
-            } else {
-                vicinity.effective_gas_price * gas_limit
+            let (gasometer_config, delete_empty) = match spec {
+                ForkSpec::Istanbul => (Config::istanbul(), true),
+                ForkSpec::Berlin => (Config::berlin(), true),
+                ForkSpec::London => (Config::london(), true),
+                ForkSpec::Merge | ForkSpec::Paris => (Config::merge(), true),
+                ForkSpec::Shanghai => (Config::shanghai(), true),
+                ForkSpec::Cancun => (Config::cancun(), true),
+                ForkSpec::Prague => (Config::prague(), true),
+                _ => {
+                    continue;
+                }
             };
 
-            // Dump state transaction data
-            let mut state_tests_dump = StateTestsDump::default();
-            state_tests_dump.set_state(&original_state);
-            state_tests_dump.set_caller_secret_key(test.unwrap_caller_secret_key());
-            state_tests_dump.set_vicinity(&vicinity);
+            // EIP-4844
+            let blob_gas_price = test
+                .0
+                .env
+                .current_excess_blob_gas
+                .map(|current_excess_blob_gas| {
+                    eip_4844::calc_blob_gas_price(current_excess_blob_gas.0.as_u64())
+                })
+                .or_else(|| {
+                    test.0
+                        .env
+                        .parent_blob_gas_used
+                        .zip(test.0.env.parent_excess_blob_gas)
+                        .map(|(parent_blob_gas_used, parent_excess_blob_gas)| {
+                            let excess_blob_gas = eip_4844::calc_excess_blob_gas(
+                                parent_blob_gas_used.0.as_u64(),
+                                parent_excess_blob_gas.0.as_u64(),
+                            );
+                            eip_4844::calc_blob_gas_price(excess_blob_gas)
+                        })
+                });
+            // EIP-4844
+            let data_max_fee = if gasometer_config.has_shard_blob_transactions {
+                let max_fee_per_blob_gas = test_tx.max_fee_per_blob_gas.unwrap_or_default().0;
+                Some(eip_4844::calc_max_data_fee(
+                    max_fee_per_blob_gas,
+                    test_tx.blob_versioned_hashes.len(),
+                ))
+            } else {
+                None
+            };
+            let data_fee = if gasometer_config.has_shard_blob_transactions {
+                Some(eip_4844::calc_data_fee(
+                    blob_gas_price.expect("expect blob_gas_price"),
+                    test_tx.blob_versioned_hashes.len(),
+                ))
+            } else {
+                None
+            };
 
-            let metadata =
-                StackSubstateMetadata::new(transaction.gas_limit.into(), &gasometer_config);
-            let executor_state = MemoryStackState::new(metadata, &backend);
-            let precompile = JsonPrecompile::precompile(spec).unwrap();
-            let mut executor =
-                StackExecutor::new_with_precompiles(executor_state, &gasometer_config, &precompile);
-            executor.state_mut().withdraw(caller, total_fee).unwrap();
-
-            let access_list: Vec<(H160, Vec<H256>)> = transaction
-                .access_list
-                .into_iter()
-                .map(|(address, keys)| (address.0, keys.into_iter().map(|k| k.0).collect()))
-                .collect();
-
-            // EIP-3607: Reject transactions from senders with deployed code
-            // EIP-7702: Accept transaction even if caller has code.
-            if caller_code.is_empty() || is_delegated {
-                match transaction.to {
-                    ethjson::maybe::MaybeEmpty::Some(to) => {
-                        let value = transaction.value.into();
-
-                        state_tests_dump.set_tx_data(
-                            to.0,
-                            value,
-                            data.clone(),
-                            gas_limit,
-                            access_list.clone(),
-                        );
-
-                        // Exit reason for Call do not analyzed as it mostly do not expect exceptions
-                        let _reason = executor.transact_call(
-                            caller,
-                            to.into(),
-                            value,
-                            data,
-                            gas_limit,
-                            access_list,
-                            authorization_list,
-                        );
-                        assert_call_exit_exception(state.expect_exception.as_ref(), name);
+            let original_state = test.unwrap_to_pre_state();
+            let vicinity = test.unwrap_to_vicinity(spec, blob_gas_price);
+            if let Err(tx_err) = vicinity {
+                tests_result.total += states.len() as u64;
+                let h = states.first().unwrap().hash.0;
+                // if vicinity could not be computed then the transaction was invalid so we simply
+                // check the original state and move on
+                let (is_valid_hash, actual_hash) = crate::utils::check_valid_hash(&h, &original_state);
+                if !is_valid_hash {
+                    tests_result.failed_tests.push(FailedTestDetails {
+                        expected_hash: h,
+                        actual_hash,
+                        index: 0,
+                        name: String::from_str(name).unwrap(),
+                        spec: spec.clone(),
+                        state: original_state,
+                    });
+                    if verbose_output.verbose_failed {
+                        println!(" [{spec:?}] {name}: {tx_err:?} ... validation failed\t<----");
                     }
-                    ethjson::maybe::MaybeEmpty::None => {
-                        let code = data;
-                        let value = transaction.value.into();
+                    tests_result.failed += 1;
+                    continue;
+                }
+                assert_vicinity_validation(&tx_err, states, spec, name, &file_name.clone());
+                // As it's expected validation error - skip the test run
+                continue;
+            }
+            let vicinity = vicinity.unwrap();
+            let caller = test.unwrap_caller();
+            let caller_balance = original_state
+                .get(&caller)
+                .map_or_else(U256::zero, |acc| acc.balance);
+            // EIP-3607
+            let caller_code = original_state
+                .get(&caller)
+                .map_or_else(Vec::new, |acc| acc.code.clone());
+            // EIP-7702 - check if it's delegated designation. If it's delegation designation then
+            // even if `caller_code` is non-empty transaction should be executed.
+            let is_delegated = original_state
+                .get(&caller)
+                .is_some_and(|c| Authorization::is_delegated(&c.code));
 
-                        let reason =
-                            executor.transact_create(caller, value, code, gas_limit, access_list);
-                        if check_create_exit_reason(
-                            &reason.0,
-                            state.expect_exception.as_ref(),
-                            &format!("{spec:?}-{name}-{i}"),
-                        ) {
-                            continue;
+            for (i, state) in states.iter().enumerate() {
+                let transaction = test_tx.select(&state.indexes);
+                let mut backend = MemoryBackend::new(&vicinity, original_state.clone());
+                tests_result.total += 1;
+                // Test case may be expected to fail with an unsupported tx type if the current fork is
+                // older than Berlin (see EIP-2718). However, this is not implemented in sputnik itself and rather
+                // in the code hosting sputnik. https://github.com/rust-blockchain/evm/pull/40
+                let expect_tx_type_not_supported = matches!(
+                    spec,
+                    ForkSpec::EIP150
+                        | ForkSpec::EIP158
+                        | ForkSpec::Frontier
+                        | ForkSpec::Homestead
+                        | ForkSpec::Byzantium
+                        | ForkSpec::Constantinople
+                        | ForkSpec::ConstantinopleFix
+                        | ForkSpec::Istanbul
+                        | ForkSpec::Berlin
+                ) && TxType::from_txbytes(&state.txbytes)
+                    != TxType::Legacy
+                    && state.expect_exception.as_deref() == Some("TR_TypeNotSupported");
+                if expect_tx_type_not_supported {
+                    continue;
+                }
+
+                let gas_limit: u64 = transaction.gas_limit.into();
+                let data: Vec<u8> = transaction.data.clone().into();
+                let valid_tx = crate::utils::transaction::validate(
+                    &transaction,
+                    test.0.env.gas_limit.0,
+                    caller_balance,
+                    &gasometer_config,
+                    test_tx,
+                    &vicinity,
+                    blob_gas_price,
+                    data_max_fee,
+                    spec,
+                    state,
+                );
+                // Only execute valid transactions
+                if let Err(err) = &valid_tx {
+                    if check_validate_exit_reason(err, state.expect_exception.as_ref(), name, spec) {
+                        continue;
+                    }
+                }
+                let authorization_list = valid_tx.unwrap();
+
+                // We do not check overflow after TX validation
+                let total_fee = if let Some(data_fee) = data_fee {
+                    vicinity.effective_gas_price * gas_limit + data_fee
+                } else {
+                    vicinity.effective_gas_price * gas_limit
+                };
+
+                // Dump state transaction data
+                let mut state_tests_dump = StateTestsDump::default();
+                state_tests_dump.set_state(&original_state);
+                state_tests_dump.set_caller_secret_key(test.unwrap_caller_secret_key());
+                state_tests_dump.set_vicinity(&vicinity);
+
+                let metadata =
+                    StackSubstateMetadata::new(transaction.gas_limit.into(), &gasometer_config);
+                let executor_state = MemoryStackState::new(metadata, &backend);
+                let precompile = JsonPrecompile::precompile(spec).unwrap();
+                let mut executor =
+                    StackExecutor::new_with_precompiles(executor_state, &gasometer_config, &precompile);
+                executor.state_mut().withdraw(caller, total_fee).unwrap();
+
+                let access_list: Vec<(H160, Vec<H256>)> = transaction
+                    .access_list
+                    .into_iter()
+                    .map(|(address, keys)| (address.0, keys.into_iter().map(|k| k.0).collect()))
+                    .collect();
+
+                // EIP-3607: Reject transactions from senders with deployed code
+                // EIP-7702: Accept transaction even if caller has code.
+                if caller_code.is_empty() || is_delegated {
+                    match transaction.to {
+                        ethjson::maybe::MaybeEmpty::Some(to) => {
+                            let value = transaction.value.into();
+
+                            state_tests_dump.set_tx_data(
+                                to.0,
+                                value,
+                                data.clone(),
+                                gas_limit,
+                                access_list.clone(),
+                            );
+
+                            // Exit reason for Call do not analyzed as it mostly do not expect exceptions
+                            let _reason = executor.transact_call(
+                                caller,
+                                to.into(),
+                                value,
+                                data,
+                                gas_limit,
+                                access_list,
+                                authorization_list,
+                            );
+                            assert_call_exit_exception(state.expect_exception.as_ref(), name);
+                        }
+                        ethjson::maybe::MaybeEmpty::None => {
+                            let code = data;
+                            let value = transaction.value.into();
+
+                            let reason =
+                                executor.transact_create(caller, value, code, gas_limit, access_list);
+                            if check_create_exit_reason(
+                                &reason.0,
+                                state.expect_exception.as_ref(),
+                                &format!("{spec:?}-{name}-{i}"),
+                            ) {
+                                continue;
+                            }
                         }
                     }
-                }
-            } else {
-                // According to EIP7702 - https://eips.ethereum.org/EIPS/eip-7702#transaction-origination:
-                // allow EOAs whose code is a valid delegation designation, i.e. `0xef0100 || address`,
-                // to continue to originate transactions.
-                #[allow(clippy::collapsible_if)]
-                if !(*spec >= ForkSpec::Prague
-                    && TxType::from_txbytes(&state.txbytes) == TxType::EOAAccountCode)
-                {
-                    assert_empty_create_caller(state.expect_exception.as_ref(), name);
-                }
-            }
-
-            let used_gas = executor.used_gas();
-            if verbose_output.print_state {
-                println!("gas_limit: {gas_limit}\nused_gas: {used_gas}");
-            }
-
-            let actual_fee = executor.fee(vicinity.effective_gas_price);
-            // Forks after London burn miner rewards and thus have different gas fee
-            // calculation (see EIP-1559)
-            let miner_reward = if spec.is_eth2() {
-                let coinbase_gas_price = vicinity
-                    .effective_gas_price
-                    .saturating_sub(vicinity.block_base_fee_per_gas);
-                executor.fee(coinbase_gas_price)
-            } else {
-                actual_fee
-            };
-
-            executor
-                .state_mut()
-                .deposit(vicinity.block_coinbase, miner_reward);
-
-            let amount_to_return_for_caller = data_fee.map_or_else(
-                || total_fee - actual_fee,
-                |data_fee| total_fee - actual_fee - data_fee,
-            );
-            executor
-                .state_mut()
-                .deposit(caller, amount_to_return_for_caller);
-
-            let (values, logs) = executor.into_state().deconstruct();
-
-            backend.apply(values, logs, delete_empty);
-            // It's special case for hard forks: London or before London
-            // According to EIP-160 empty account should be removed. But in that particular test - original test state
-            // contains account 0x03 (it's precompile), and when precompile 0x03 was called it exit with
-            // OutOfGas result. And after exit of substate account not marked as touched, as exit reason
-            // is not success. And it mean, that it don't appeared in Apply::Modify, then as untouched it
-            // can't be removed by backend.apply event. In that particular case we should manage it manually.
-            // NOTE: it's not realistic situation for real life flow.
-            if *spec <= ForkSpec::London && delete_empty && name == "failed_tx_xcf416c53" {
-                let state = backend.state_mut();
-                state.retain(|addr, account| {
-                    // Check is account empty for precompile 0x03
-                    !(addr == &H160::from_low_u64_be(3)
-                        && account.balance == U256::zero()
-                        && account.nonce == U256::zero()
-                        && account.code.is_empty())
-                });
-            }
-
-            let (is_valid_hash, actual_hash) =
-                crate::utils::check_valid_hash(&state.hash.0, backend.state());
-
-            if !is_valid_hash {
-                let failed_res = FailedTestDetails {
-                    expected_hash: state.hash.0,
-                    actual_hash,
-                    index: i,
-                    name: String::from_str(name).unwrap(),
-                    spec: spec.clone(),
-                    state: backend.state().clone(),
-                };
-                tests_result.failed_tests.push(failed_res);
-                tests_result.failed += 1;
-
-                if verbose_output.verbose_failed {
-                    println!("\n[{spec:?}] {name}:{i} ... failed\t<----");
+                } else {
+                    // According to EIP7702 - https://eips.ethereum.org/EIPS/eip-7702#transaction-origination:
+                    // allow EOAs whose code is a valid delegation designation, i.e. `0xef0100 || address`,
+                    // to continue to originate transactions.
+                    #[allow(clippy::collapsible_if)]
+                    if !(*spec >= ForkSpec::Prague
+                        && TxType::from_txbytes(&state.txbytes) == TxType::EOAAccountCode)
+                    {
+                        assert_empty_create_caller(state.expect_exception.as_ref(), name);
+                    }
                 }
 
+                let used_gas = executor.used_gas();
                 if verbose_output.print_state {
-                    // Print detailed state data
-                    println!(
-                        "expected_hash:\t{:?}\nactual_hash:\t{actual_hash:?}",
-                        state.hash.0,
-                    );
-                    for (addr, acc) in backend.state().clone() {
-                        // Decode balance
-                        let balance = acc.balance.to_string();
-
-                        println!(
-                            "{addr:?}: {{\n    balance: {balance}\n    code: {:?}\n    nonce: {}\n    storage: {:#?}\n}}",
-                            hex::encode(acc.code),
-                            acc.nonce,
-                            acc.storage
-                        );
-                    }
-                    if let Some(e) = state.expect_exception.as_ref() {
-                        println!("-> expect_exception: {e}");
-                    }
+                    println!("gas_limit: {gas_limit}\nused_gas: {used_gas}");
                 }
-            } else if verbose_output.very_verbose && !verbose_output.verbose_failed {
-                println!(" [{spec:?}]  {name}:{i} ... passed");
-            }
 
-            state_tests_dump.set_used_gas(used_gas);
-            state_tests_dump.set_state_hash(actual_hash);
-            state_tests_dump.set_result_state(backend.state());
-            state_tests_dump.dump_to_file(spec);
+                let actual_fee = executor.fee(vicinity.effective_gas_price);
+                // Forks after London burn miner rewards and thus have different gas fee
+                // calculation (see EIP-1559)
+                let miner_reward = if spec.is_eth2() {
+                    let coinbase_gas_price = vicinity
+                        .effective_gas_price
+                        .saturating_sub(vicinity.block_base_fee_per_gas);
+                    executor.fee(coinbase_gas_price)
+                } else {
+                    actual_fee
+                };
+
+                executor
+                    .state_mut()
+                    .deposit(vicinity.block_coinbase, miner_reward);
+
+                let amount_to_return_for_caller = data_fee.map_or_else(
+                    || total_fee - actual_fee,
+                    |data_fee| total_fee - actual_fee - data_fee,
+                );
+                executor
+                    .state_mut()
+                    .deposit(caller, amount_to_return_for_caller);
+
+                let (values, logs) = executor.into_state().deconstruct();
+
+                backend.apply(values, logs, delete_empty);
+                // It's special case for hard forks: London or before London
+                // According to EIP-160 empty account should be removed. But in that particular test - original test state
+                // contains account 0x03 (it's precompile), and when precompile 0x03 was called it exit with
+                // OutOfGas result. And after exit of substate account not marked as touched, as exit reason
+                // is not success. And it mean, that it don't appeared in Apply::Modify, then as untouched it
+                // can't be removed by backend.apply event. In that particular case we should manage it manually.
+                // NOTE: it's not realistic situation for real life flow.
+                if *spec <= ForkSpec::London && delete_empty && name == "failed_tx_xcf416c53" {
+                    let state = backend.state_mut();
+                    state.retain(|addr, account| {
+                        // Check is account empty for precompile 0x03
+                        !(addr == &H160::from_low_u64_be(3)
+                            && account.balance == U256::zero()
+                            && account.nonce == U256::zero()
+                            && account.code.is_empty())
+                    });
+                }
+
+                let (is_valid_hash, actual_hash) =
+                    crate::utils::check_valid_hash(&state.hash.0, backend.state());
+
+                if !is_valid_hash {
+                    let failed_res = FailedTestDetails {
+                        expected_hash: state.hash.0,
+                        actual_hash,
+                        index: i,
+                        name: String::from_str(name).unwrap(),
+                        spec: spec.clone(),
+                        state: backend.state().clone(),
+                    };
+                    tests_result.failed_tests.push(failed_res);
+                    tests_result.failed += 1;
+
+                    if verbose_output.verbose_failed {
+                        println!("\n[{spec:?}] {name}:{i} ... failed\t<----");
+                    }
+
+                    if verbose_output.print_state {
+                        // Print detailed state data
+                        println!(
+                            "expected_hash:\t{:?}\nactual_hash:\t{actual_hash:?}",
+                            state.hash.0,
+                        );
+                        for (addr, acc) in backend.state().clone() {
+                            // Decode balance
+                            let balance = acc.balance.to_string();
+
+                            println!(
+                                "{addr:?}: {{\n    balance: {balance}\n    code: {:?}\n    nonce: {}\n    storage: {:#?}\n}}",
+                                hex::encode(acc.code),
+                                acc.nonce,
+                                acc.storage
+                            );
+                        }
+                        if let Some(e) = state.expect_exception.as_ref() {
+                            println!("-> expect_exception: {e}");
+                        }
+                    }
+                } else if verbose_output.very_verbose && !verbose_output.verbose_failed {
+                    println!(" [{spec:?}]  {name}:{i} ... passed");
+                }
+
+                state_tests_dump.set_used_gas(used_gas);
+                state_tests_dump.set_state_hash(actual_hash);
+                state_tests_dump.set_result_state(backend.state());
+                state_tests_dump.dump_to_file(spec);
+            }
         }
-    }
+    */
     tests_result
 }
